fix(condo): DOMA-7709 moved InvoiceRowsTable to own file
